+++ CHANGELOG.md
@@ -17,6 +17,8 @@ These are the section headers that we use:
 
 ## [Unreleased]
 
+- Added `ArgillaAutoTrainTrainer` for Text Classification [#2664](https://github.com/argilla-io/argilla/issues/2664)
+
 ## [1.7.0](https://github.com/argilla-io/argilla/compare/v1.6.0...v1.7.0)
 
 ### Added